Add files via upload
This is my personal portfolio website built using HTML, CSS, and JavaScript to showcase my skills, projects, and experience as a web developer. The website includes the following sections:

Home: Introduction and background.
Projects: A showcase of my latest web development projects.
About: Information about my skills, education, and experience.
Contact: A form to get in touch with me.
The site is fully responsive and optimized for various devices.